Back to 624 bytes
I've managed to get the code back to 13 lines (and 624 bytes) like it was while including the Windows fix. However, as a consequence, this places your cursor at the start of the text if you quit meaning you have to clear the terminal or  type through the text. But this is worth it for the Windows fix and removing the unnecessary line. The comments are also gone now.